+++ scriptabit/tests/test_utility_functions.py
@@ -38,25 +38,106 @@ class TestUtilityFunctions(object):
             m.put('https://habitica.com/api/v3/user',
                   text=get_fake_stats(hp=39)[1])
             uf = UtilityFunctions(MockConfig(), self.hs)
-            assert uf.set_health(39) == 39
+            uf.set_health(39)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.hp=39'
+
+    def test_inc_health(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(hp=1)[1])
+            m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(hp=8)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_health(7, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.hp=8'
+
+    def test_dec_health(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(hp=50)[1])
+            m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(hp=29)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_health(-21, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.hp=29'
 
     def test_set_mana(self):
         with requests_mock.mock() as m:
             m.get('https://habitica.com/api/v3/user',
                   text=get_fake_stats()[1])
             m.put('https://habitica.com/api/v3/user',
-                  text=get_fake_stats(mp=9)[1])
+                  text=get_fake_stats(mp=100)[1])
             uf = UtilityFunctions(MockConfig(), self.hs)
-            assert uf.set_mana(9) == 9
+            uf.set_mana(100)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.mp=100'
+
+    def test_inc_mana(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(mp=30)[1])
+            m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(mp=99)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_mana(69, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.mp=99'
+
+    def test_dec_mana(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(mp=50)[1])
+            m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(mp=39)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_mana(-11, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.mp=39'
 
     def test_set_xp(self):
         with requests_mock.mock() as m:
             m.get('https://habitica.com/api/v3/user',
                   text=get_fake_stats()[1])
             m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(exp=1009)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_xp(1009)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.exp=1009'
+
+    def test_inc_xp(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(exp=30)[1])
+            m.put('https://habitica.com/api/v3/user',
                   text=get_fake_stats(exp=39)[1])
             uf = UtilityFunctions(MockConfig(), self.hs)
-            assert uf.set_xp(39) == 39
+            uf.set_xp(9, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.exp=39'
+
+    def test_dec_xp(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(exp=500)[1])
+            m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(exp=120)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_xp(-380, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.exp=120'
 
     def test_set_gold(self):
         with requests_mock.mock() as m:
@@ -65,7 +146,34 @@ class TestUtilityFunctions(object):
             m.put('https://habitica.com/api/v3/user',
                   text=get_fake_stats(gp=9009)[1])
             uf = UtilityFunctions(MockConfig(), self.hs)
-            assert uf.set_gold(9009) == 9009
+            uf.set_gold(9009)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.gp=9009'
+
+    def test_inc_gold(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(gp=30)[1])
+            m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(gp=39)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_gold(9, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.gp=39'
+
+    def test_dec_gold(self):
+        with requests_mock.mock() as m:
+            m.get('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(gp=50)[1])
+            m.put('https://habitica.com/api/v3/user',
+                  text=get_fake_stats(gp=39)[1])
+            uf = UtilityFunctions(MockConfig(), self.hs)
+            uf.set_gold(-11, True)
+            history = m.request_history
+            assert history[1].method == 'PUT'
+            assert history[1].text == 'stats.gp=39'
 
     def test_set_health_dry_run(self):
         with requests_mock.mock() as m: